backend/requirements.txt
azure-identity==1.14.0
azure-search-documents==11.4.0b6
azure-storage-blob==12.19.1
quart==0.19.3
uvicorn[standard]==0.23.2
openai==1.6.1
pandas==2.2.1